% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{addCategoricalAttribute}
\alias{addCategoricalAttribute}
\title{adds a new query to the plot}
\usage{
addCategoricalAttribute(upsetjs, name, values, categories = NULL)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{name}{name of the attribute}

\item{values}{the values as a factor}

\item{categories}{optional categories otherweise the levels are used}
}
\value{
the object given as first argument
}
\description{
adds a new query to the plot
}
\examples{
upsetjs() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  addCategoricalAttribute("attr", as.factor(sample(c("male", "female"), 3, replace = TRUE)))
}
