% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{addNumericAttribute}
\alias{addNumericAttribute}
\title{adds a new numeric attribute to the plot}
\usage{
addNumericAttribute(upsetjs, name, values, min_value = NULL, max_value = NULL)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{name}{name of the attribute}

\item{values}{the values as a numeric vector}

\item{min_value}{optional min domain value}

\item{max_value}{optional max domain value}
}
\value{
the object given as first argument
}
\description{
adds a new numeric attribute to the plot
}
\examples{
upsetjs() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  addNumericAttribute("attr", runif(3))
}
