% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{addQuery}
\alias{addQuery}
\title{adds a new query to the plot}
\usage{
addQuery(upsetjs, name, color, elems = NULL, set = NULL)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{name}{name of the query}

\item{color}{color of the query}

\item{elems}{the list of elems to highlight}

\item{set}{the set name, similar to the selection}
}
\value{
the object given as first argument
}
\description{
adds a new query to the plot
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% addQuery(name="Q1", color="red", set="b")

}
