% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{asSet}
\alias{asSet}
\title{creates a new UpSet set structure}
\usage{
asSet(name, elems = c(), cardinality = length(elems), color = NULL)
}
\arguments{
\item{name}{name of the set}

\item{elems}{the elements of the set}

\item{cardinality}{the cardinality of the set, default to \code{length(elems)}}

\item{color}{the color of the set}
}
\value{
the set object
}
\description{
creates a new UpSet set structure
}
\examples{
asSet("a", c(1, 2, 3))
}
