% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartFontSizes}
\alias{chartFontSizes}
\title{specify chart font sizes}
\usage{
chartFontSizes(
  upsetjs,
  font.family = NULL,
  chart.label = NULL,
  set.label = NULL,
  axis.tick = NULL,
  bar.label = NULL,
  legend = NULL
)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{font.family}{specify the font family to render}

\item{chart.label}{font size of the chart label, default: 16px}

\item{set.label}{font size of the set label, default: 10px}

\item{axis.tick}{font size of the axis tick, default: 16px}

\item{bar.label}{font size of the bar label, default: 10px}

\item{legend}{font size of the legend label, default: 10px}
}
\value{
the object given as first argument
}
\description{
specify chart font sizes
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% chartFontSizes(font.family="serif")

}
