% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartKarnaughMapLabels}
\alias{chartKarnaughMapLabels}
\title{specify chart labels}
\usage{
chartKarnaughMapLabels(upsetjs, title = NULL, description = NULL)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs_kamp} or \code{upsetjs_kmap_proxy}}

\item{title}{the chart title}

\item{description}{the chart description}
}
\value{
the object given as first argument
}
\description{
specify chart labels
}
\examples{
upsetjsKarnaughMap() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  chartKarnaughMapLabels(title = "Test")
}
