% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartKarnaughMapLayout}
\alias{chartKarnaughMapLayout}
\title{specify the chart karnaugh map layout}
\usage{
chartKarnaughMapLayout(
  upsetjs,
  padding = NULL,
  bar.padding = NULL,
  numerical.scale = NULL
)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs_kmap} or \code{upsetjs_kmap_proxy}}

\item{padding}{padding around the plot}

\item{bar.padding}{padding ratio (default 0.1) for the bar charts}

\item{numerical.scale}{numerical scale: linear (default) or log}
}
\value{
the object given as first argument
}
\description{
specify the chart karnaugh map layout
}
\examples{
upsetjsKarnaughMap() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% chartKarnaughMapLayout(padding=10)

}
