% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartLabels}
\alias{chartLabels}
\title{specify chart labels}
\usage{
chartLabels(
  upsetjs,
  combination.name = NULL,
  combination.name.axis.offset = NULL,
  set.name = NULL,
  set.name.axis.offset = NULL,
  bar.label.offset = NULL
)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{combination.name}{the label for the combination chart}

\item{combination.name.axis.offset}{the offset of the combination label from the axis in pixel}

\item{set.name}{the label for the set chart}

\item{set.name.axis.offset}{the offset of the set label from the axis in pixel}

\item{bar.label.offset}{the offset of the bar label from the bar in pixel}
}
\value{
the object given as first argument
}
\description{
specify chart labels
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% chartLabels(set.name="Test")

}
