% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartProps}
\alias{chartProps}
\title{generic set chart props}
\usage{
chartProps(upsetjs, ...)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{...}{all upsetjs properties in R name notation}
}
\value{
the object given as first argument
}
\description{
generic set chart props
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% chartProps(theme="dark")

}
