% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartStyleFlags}
\alias{chartStyleFlags}
\title{specify chart flags}
\usage{
chartStyleFlags(upsetjs, id = NULL, export.buttons = NULL, class.name = NULL)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{id}{the optional HTML ID}

\item{export.buttons}{show export SVG and PNG buttons}

\item{class.name}{extra CSS class name to the root element}
}
\value{
the object given as first argument
}
\description{
specify chart flags
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% chartStyleFlags(id="test")

}
