% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartTheme}
\alias{chartTheme}
\title{specify theming options}
\usage{
chartTheme(
  upsetjs,
  theme = NULL,
  selection.color = NULL,
  alternating.color = NULL,
  color = NULL,
  has.selection.color = NULL,
  text.color = NULL,
  hover.hint.color = NULL,
  not.member.color = NULL,
  value.text.color = NULL,
  stroke.color = NULL,
  has.selection.opacity = NULL,
  opacity = NULL,
  filled = NULL
)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{theme}{theme to use 'dark' or 'light'}

\item{selection.color}{selection color}

\item{alternating.color}{alternating background color}

\item{color}{main bar color}

\item{has.selection.color}{main color used when a selection is present}

\item{text.color}{main text color}

\item{hover.hint.color}{color of the hover hint}

\item{not.member.color}{color of the dot if not a member}

\item{value.text.color}{value text color (venn diagram only)}

\item{stroke.color}{circle stroke color (venn diagram and karnaugh map only)}

\item{has.selection.opacity}{main opacity used when a selection is present}

\item{opacity}{main bar opacity}

\item{filled}{enforce filled circles (venn diagram only)}
}
\value{
the object given as first argument
}
\description{
specify theming options
}
\examples{
upsetjs() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  chartTheme(theme = "dark")
}
