% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartVennLabels}
\alias{chartVennLabels}
\title{specify chart labels}
\usage{
chartVennLabels(upsetjs, title = NULL, description = NULL)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs_venn} or \code{upsetjs_venn_proxy}}

\item{title}{the chart title}

\item{description}{the chart description}
}
\value{
the object given as first argument
}
\description{
specify chart labels
}
\examples{
upsetjsVennDiagram() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% chartVennLabels(title="Test")

}
