% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{chartVennLayout}
\alias{chartVennLayout}
\title{specify the chart venn layout}
\usage{
chartVennLayout(upsetjs, padding = NULL)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs_venn} or \code{upsetjs_venn_proxy}}

\item{padding}{padding around the plot}
}
\value{
the object given as first argument
}
\description{
specify the chart venn layout
}
\examples{
upsetjsVennDiagram() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  chartVennLayout(padding = 10)
}
