% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{clearAttributes}
\alias{clearAttributes}
\title{clears the list of attributes for incremental updates}
\usage{
clearAttributes(upsetjs)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}
}
\value{
the object given as first argument
}
\description{
clears the list of attributes for incremental updates
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\%
clearAttributes()

}
