% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{clearQueries}
\alias{clearQueries}
\title{clears the list of queries for incremental updates}
\usage{
clearQueries(upsetjs)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}
}
\value{
the object given as first argument
}
\description{
clears the list of queries for incremental updates
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\%
addQuery(name="Q1", color="red", set="b") \%>\% clearQueries()

}
