% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fromList}
\alias{fromList}
\title{generates the sets from a lists object}
\usage{
fromList(
  upsetjs,
  value,
  order.by = "cardinality",
  limit = NULL,
  shared = NULL,
  shared.mode = "click"
)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{value}{the list input value}

\item{order.by}{order intersections by cardinality or name}

\item{limit}{limit the ordered sets to the given limit}

\item{shared}{a crosstalk shared data frame}

\item{shared.mode}{whether on 'hover' or 'click' (default) is synced}
}
\value{
the object given as first argument
}
\description{
generates the sets from a lists object
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3)))

}
