% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generateDistinctIntersections}
\alias{generateDistinctIntersections}
\title{configure the generation of the distinct intersections}
\usage{
generateDistinctIntersections(
  upsetjs,
  min = 0,
  max = NULL,
  empty = FALSE,
  order.by = "cardinality",
  limit = NULL,
  colors = NULL
)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{min}{minimum number of sets in an intersection}

\item{max}{maximum number of sets in an intersection}

\item{empty}{whether to include empty intersections or not}

\item{order.by}{order intersections by cardinality, degree, name or a combination of it}

\item{limit}{limit the number of intersections to the top N}

\item{colors}{the optional list with set name to color}
}
\value{
the object given as first argument
}
\description{
configure the generation of the distinct intersections
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% generateDistinctIntersections(min=2)

}
