% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getCombinations}
\alias{getCombinations}
\title{extract the vector of combinations}
\usage{
getCombinations(upsetjs)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs}}
}
\value{
vector of sets
}
\description{
extract the vector of combinations
}
\examples{
upsetjs() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  getCombinations()
}
