% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getElements}
\alias{getElements}
\title{extract the vector of elements}
\usage{
getElements(upsetjs)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}
}
\value{
vector of elements
}
\description{
extract the vector of elements
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\% getElements()

}
