% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getSets}
\alias{getSets}
\title{extract the vector of sets}
\usage{
getSets(upsetjs)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs}}
}
\value{
vector of sets
}
\description{
extract the vector of sets
}
\examples{
upsetjs() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  getSets()
}
