% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got.R
\docType{data}
\name{got}
\alias{got}
\title{Games of Thrones Character dataset for UpSet.js}
\format{A data frame with 22 rows and 6 variables/sets:
\describe{
\item{Lannister}{character part of the Lannister house}
\item{Stark}{character part of the Start house}
\item{female}{character is female}
\item{male}{character is male}
\item{royal}{character is royal}
\item{was.killed}{character was killed}
}}
\source{
\url{https://github.com/jeffreylancaster/game-of-thrones}
}
\usage{
got
}
\description{
A dataset containing set information about Game of Thrones characters
}
\keyword{datasets}
