% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{interactiveChart}
\alias{interactiveChart}
\title{make it an interactive chart}
\usage{
interactiveChart(upsetjs, value = TRUE, events_nonce = FALSE)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{value}{whether to enable or disable or set the mode: hover, click, contextMenu}

\item{events_nonce}{whether to enable send a unique once (event date) for each event to prevent deduplication}
}
\value{
the object given as first argument
}
\description{
make it an interactive chart
}
\examples{
upsetjs() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  interactiveChart()
}
