% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryLegend}
\alias{queryLegend}
\title{renders a legend for the queries}
\usage{
queryLegend(upsetjs, value = TRUE)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{value}{whether to enable or disable}
}
\value{
the object given as first argument
}
\description{
renders a legend for the queries
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\%
addQuery(name="Q1", color="red", set="b") \%>\% queryLegend(FALSE)

}
