% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\title{reexport operator}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

