% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{setAttributes}
\alias{setAttributes}
\title{set the attributes}
\usage{
setAttributes(upsetjs, attrs = list())
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{attrs}{the attributes to set}
}
\value{
the object given as first argument
}
\description{
set the attributes
}
\examples{
upsetjs() \%>\% fromList(list(a=c(1,2,3), b=c(2,3))) \%>\%
setAttributes(list(attr=runif(3), cat=as.factor(sample(c('male','female'), 3, replace=TRUE))))

}
