% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setCombinations}
\alias{setCombinations}
\title{set the vector of combinations}
\usage{
setCombinations(upsetjs, value)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs}}

\item{value}{the vector of combinations}
}
\value{
the object given as first argument
}
\description{
set the vector of combinations
}
\examples{
upsetjs() \%>\%
  setCombinations(list(asCombination("a", c(1, 2, 3)))) \%>\%
  getCombinations()
}
