% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{setQueries}
\alias{setQueries}
\title{set the queries}
\usage{
setQueries(upsetjs, queries = list())
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{queries}{the queries to set}
}
\value{
the object given as first argument
}
\description{
set the queries
}
\examples{
upsetjs() \%>\%
  fromList(list(a = c(1, 2, 3), b = c(2, 3))) \%>\%
  setQueries(list(list(name = "Q1", color = "red", set = "b")))
}
