% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{setSets}
\alias{setSets}
\title{set the vector of sets}
\usage{
setSets(upsetjs, value)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs}}

\item{value}{the vector of sets}
}
\value{
the object given as first argument
}
\description{
set the vector of sets
}
\examples{
upsetjs() \%>\% setCombinations(list(asSet('a', c(1,2,3)))) \%>\% getSets()

}
