% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{upsetjsDash}
\alias{upsetjsDash}
\title{create a new upsetjs dash adapter}
\usage{
upsetjsDash(children = NULL, id = NULL, width = NULL, height = NULL)
}
\arguments{
\item{children}{dash children}

\item{id}{dash id}

\item{width}{upsetjs width}

\item{height}{upestjs height}
}
\value{
the set object
}
\description{
create a new upsetjs dash adapter
}
\examples{
upsetjsDash("u") \%>\% fromList(list(a = c(1, 2, 3), b = c(2, 3)))
}
