% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs.R
\name{upsetjsEulerDiagramProxy}
\alias{upsetjsEulerDiagramProxy}
\title{reactive helper to update an upsetjs euler diagram in place}
\usage{
upsetjsEulerDiagramProxy(outputId, session)
}
\arguments{
\item{outputId}{the id of the upsetjs widget}

\item{session}{current shiny session}
}
\value{
an object of class \code{upsetjs_proxy}
}
\description{
reactive helper to update an upsetjs euler diagram in place
}
\examples{
\dontrun{
upsetjsEulerDiagramProxy('upsetjs1', session) \%>\% setSelection('a')
}
}
