% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs.R
\name{upsetjsKarnaughMap}
\alias{upsetjsKarnaughMap}
\title{upsetjs - factory for UpSet.js Karnaugh Map HTMLWidget}
\usage{
upsetjsKarnaughMap(
  width = "100\%",
  height = NULL,
  elementId = NULL,
  sizingPolicy = upsetjsSizingPolicy()
)
}
\arguments{
\item{width}{width of the element}

\item{height}{height of the element}

\item{elementId}{unique element id}

\item{sizingPolicy}{htmlwidgets sizing policy object. Defaults to \code{\link{upsetjsSizingPolicy}()}}
}
\value{
An object of class \code{upsetjs_venn} and \code{htmlwidget}
}
\description{
upsetjs - factory for UpSet.js Karnaugh Map HTMLWidget
}
\examples{
upsetjsKarnaughMap() \%>\% fromList(list(a = c(1, 2, 3), b = c(2, 3)))
}
