% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs.R
\name{upsetjsKarnaughMapProxy}
\alias{upsetjsKarnaughMapProxy}
\title{reactive helper to update an upsetjs karnaugh map diagram in place}
\usage{
upsetjsKarnaughMapProxy(outputId, session)
}
\arguments{
\item{outputId}{the id of the upsetjs widget}

\item{session}{current shiny session}
}
\value{
an object of class \code{upsetjs_proxy}
}
\description{
reactive helper to update an upsetjs karnaugh map diagram in place
}
\examples{
\dontrun{
upsetjsKarnaughMapProxy("upsetjs1", session) \%>\% setSelection("a")
}
}
