% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetjs.R
\name{upsetjsVennDiagramProxy}
\alias{upsetjsVennDiagramProxy}
\title{reactive helper to update an upsetjs venn diagram in place}
\usage{
upsetjsVennDiagramProxy(outputId, session)
}
\arguments{
\item{outputId}{the id of the upsetjs widget}

\item{session}{current shiny session}
}
\value{
an object of class \code{upsetjs_proxy}
}
\description{
reactive helper to update an upsetjs venn diagram in place
}
\examples{
\dontrun{
upsetjsVennDiagramProxy("upsetjs1", session) \%>\% setSelection("a")
}
}
