% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fromExpression}
\alias{fromExpression}
\title{generates the sets from a lists object that contained the cardinalities of both sets and combinations (&)}
\usage{
fromExpression(
  upsetjs,
  value,
  symbol = "&",
  order.by = "cardinality",
  colors = NULL,
  type = "intersection"
)
}
\arguments{
\item{upsetjs}{an object of class \code{upsetjs} or \code{upsetjs_proxy}}

\item{value}{the expression list input}

\item{symbol}{the symbol how to split list names to get the sets}

\item{order.by}{order intersections by cardinality or name}

\item{colors}{the optional list with set name to color}

\item{type}{the type of intersections this data represents (intersection,union,distinctIntersection)}
}
\value{
the object given as first argument
}
\description{
generates the sets from a lists object that contained the cardinalities of both sets and combinations (&)
}
\examples{
upsetjs() \%>\% fromExpression(list(a = 3, b = 2, `a&b` = 2))
}
